Use standard code formatting

+++ src/transformers/modeling_prophetnet.py
@@ -583,9 +583,7 @@ class ProphetNetSelfAttention(nn.Module):
     """Multi-headed attention from 'Attention Is All You Need' paper"""
 
     def __init__(
-        self,
-        config: ProphetNetConfig,
-        num_attn_heads: int,
+        self, config: ProphetNetConfig, num_attn_heads: int,
     ):
         super().__init__()
         hidden_size = config.hidden_size
@@ -679,11 +677,7 @@ class ProphetNetSelfAttention(nn.Module):
             attn_weights = attn_weights + attention_mask
 
         attn_weights = F.softmax(attn_weights, dim=-1)
-        attn_probs = F.dropout(
-            attn_weights,
-            p=self.attention_dropout,
-            training=self.training,
-        )
+        attn_probs = F.dropout(attn_weights, p=self.attention_dropout, training=self.training,)
 
         attn_output = torch.bmm(attn_probs, value_states)
         assert attn_output.size() == (
@@ -838,11 +832,7 @@ class ProphetNetNgramProphetNetSelfAttention(nn.Module):
         if attention_mask is not None:
             main_attn_weights = main_attn_weights + attention_mask
 
-        main_attn_probs = softmax(
-            main_attn_weights,
-            dim=-1,
-            onnx_trace=self.onnx_trace,
-        ).type_as(main_attn_weights)
+        main_attn_probs = softmax(main_attn_weights, dim=-1, onnx_trace=self.onnx_trace,).type_as(main_attn_weights)
 
         main_attn_probs = F.dropout(main_attn_probs, p=self.attention_dropout, training=self.training)
 
@@ -887,11 +877,9 @@ class ProphetNetNgramProphetNetSelfAttention(nn.Module):
         if extended_predict_attention_mask is not None:
             predict_attn_weights = predict_attn_weights + extended_predict_attention_mask
 
-        predict_attn_probs = softmax(
-            predict_attn_weights,
-            dim=-1,
-            onnx_trace=self.onnx_trace,
-        ).type_as(predict_attn_weights)
+        predict_attn_probs = softmax(predict_attn_weights, dim=-1, onnx_trace=self.onnx_trace,).type_as(
+            predict_attn_weights
+        )
         predict_attn_probs = F.dropout(predict_attn_probs, p=self.attention_dropout, training=self.training)
 
         # project to attention output
@@ -1029,10 +1017,7 @@ class ProphetNetEncoderLayer(nn.Module):
 
     def forward(self, hidden_states, attention_mask):
         # 1st residual block
-        attention_output, attn_weights = self.self_attn(
-            hidden_states=hidden_states,
-            attention_mask=attention_mask,
-        )
+        attention_output, attn_weights = self.self_attn(hidden_states=hidden_states, attention_mask=attention_mask,)
         hidden_states = self.self_attn_layer_norm(attention_output + hidden_states)
 
         # 2nd residual block
@@ -1112,8 +1097,7 @@ class ProphetNetDecoderLayer(nn.Module):
 
 
 @add_start_docstrings(
-    "The standalone encoder part of the ProphetNetModel.",
-    PROPHETNET_START_DOCSTRING,
+    "The standalone encoder part of the ProphetNetModel.", PROPHETNET_START_DOCSTRING,
 )
 class ProphetNetEncoder(ProphetNetPreTrainedModel):
     r"""
@@ -1221,8 +1205,7 @@ class ProphetNetEncoder(ProphetNetPreTrainedModel):
 
 
 @add_start_docstrings(
-    "The standalone decoder part of the ProphetNetModel.",
-    PROPHETNET_START_DOCSTRING,
+    "The standalone decoder part of the ProphetNetModel.", PROPHETNET_START_DOCSTRING,
 )
 class ProphetNetDecoder(ProphetNetPreTrainedModel):
     r"""
@@ -1326,9 +1309,7 @@ class ProphetNetDecoder(ProphetNetPreTrainedModel):
         batch_size, sequence_length = inputs_embeds.shape[:2]
 
         main_stream_pos_embed, position_ids = self.position_embeddings(
-            (batch_size, sequence_length),
-            device=inputs_embeds.device,
-            past_key_values=past_key_values,
+            (batch_size, sequence_length), device=inputs_embeds.device, past_key_values=past_key_values,
         )
 
         if past_key_values is not None:
@@ -1780,11 +1761,7 @@ class ProphetNetForConditionalGeneration(ProphetNetPreTrainedModel):
                 break
             expend_targets[i, :, :] = labels
 
-        lprobs = F.log_softmax(
-            logits.view(-1, logits.size(-1)),
-            dim=-1,
-            dtype=torch.float32,
-        )
+        lprobs = F.log_softmax(logits.view(-1, logits.size(-1)), dim=-1, dtype=torch.float32,)
 
         loss = F.nll_loss(lprobs, expend_targets.view(-1), reduction="sum")
 
@@ -1996,11 +1973,7 @@ class ProphetNetForCausalLM(ProphetNetPreTrainedModel):
                 break
             expend_targets[i, :, :] = labels
 
-        lprobs = F.log_softmax(
-            logits.view(-1, logits.size(-1)),
-            dim=-1,
-            dtype=torch.float32,
-        )
+        lprobs = F.log_softmax(logits.view(-1, logits.size(-1)), dim=-1, dtype=torch.float32,)
 
         loss = F.nll_loss(lprobs, expend_targets.view(-1), reduction="sum")
